fixed Device to match data type for Dev size on Darwin
This will break on Linux.

+++ event.go
@@ -16,5 +16,5 @@ type FileState struct {
   Source *string `json:"source,omitempty"`
   Offset int64 `json:"offset,omitempty"`
   Inode uint64 `json:"inode,omitempty"`
-  Device uint64 `json:"device,omitempty"`
+  Device int32 `json:"device,omitempty"`
 }
